progress with accordion, dynamic icons now
