+++ src/components/Accordion.css
@@ -38,8 +38,7 @@
 
 .Accordion .itemcontent {
   display: none;
-  background: white;
-  color: #9e8eb9; }
+  background: white; }
 
 .Accordion .itemcontent ul {
   padding: 6.25rem 10rem 0 10rem;
@@ -47,20 +46,21 @@
   margin: 0; }
 
 .Accordion .itemcontent ul li {
-  color: #9e8eb9;
-  border-bottom: 3px solid #9e8eb9;
+  text-align: center;
   border-bottom-width: 3px;
-  border-bottom-style: solid;
-  border-bottom-color: #9e8eb9;
-  text-align: center; }
+  border-bottom-style: solid; }
 
-.Accordion .itemcontent ul li .imageholder {
+.Accordion .itemcontent ul li.selected {
+  color: #9e8eb9;
+  border-bottom-color: #9e8eb9; }
+
+.Accordion .itemcontent ul li.selected .imageholder {
   border: 3px solid #9e8eb9;
   display: inline-block;
   height: 80px;
   border-radius: 27px; }
 
-.Accordion .itemcontent ul img {
+.Accordion .itemcontent ul li.selected .imageholder img {
   filter: invert(61%) sepia(4%) saturate(2427%) hue-rotate(220deg) brightness(96%) contrast(87%); }
 
 .Accordion .itemcontent.open {
