+++ src/components/Accordion.js
@@ -1,18 +1,18 @@
 import './Accordion.css';
 import React from 'react';
 
-export default function Accordion({data}) {
+export default function Accordion({data,accordiondata}) {
 
     function openaccordion(id) {
         let elem = document.getElementById(id);
         elem.classList.contains("open") ? elem.classList.remove("open") : elem.classList.add("open");
     }
-
+    const {iconForFeature} = accordiondata;
     return(        
         <div className="Accordion">
             {
                 data.map((entry,i)  => {
-                    const {wpParent , name /*,wpChildren*/} = entry;
+                    const {wpParent , name ,wpChildren} = entry;
                     if (wpParent === null) {
                         return (     
                             <div key={`accordion-${name}`}>  
@@ -22,12 +22,21 @@ export default function Accordion({data}) {
                                 </div>
                                 <div className="itemcontent" id={`accordion-${name}`}>
                                     <ul>
-                                        <li>
-                                            <div className='imageholder'>
-                                                <img width="80" height="80" src='http://localhost:8000/wp-content/uploads/2022/04/Create-Contour.png' alt=""/>
-                                            </div>
-                                            <h5>{name}</h5>                                     
-                                        </li>
+                                        {
+                                            wpChildren.nodes.map((child) => {
+
+                                                let feature = iconForFeature.find((elem) => elem.feature.name === child.name);
+                                                let url = feature.icon.sourceUrl;
+                                                console.log(feature);
+
+                                                return (<li>
+                                                    <div className='imageholder'>
+                                                        <img width="80" height="80" src='http://localhost:8000/wp-content/uploads/2022/04/Create-Contour.png' alt=""/>
+                                                    </div>
+                                                    <h5>{child.name}</h5>                                     
+                                                </li>);
+                                            })
+                                        }                                        
                                     </ul>
                                     
                                 </div>
